requirements.txt
streamlit
pandas
numpy
folium
streamlit-folium
matplotlib
pillow
torch 